edit the comments and edit ui
