+++ lib/core/extension/num_extension.dart
@@ -1,11 +1,11 @@
 import '../services/responsive/responsive.dart';
 
 extension SizeExtension on num {
-  double get w => ResponsiveService.scaleWidth() * this;
+  double get width => ResponsiveService.scaleWidth() * this;
 
-  double get h => ResponsiveService.scaleHeight() * this;
+  double get height => ResponsiveService.scaleHeight() * this;
 
-  double get r => ResponsiveService.scaleRadius() * this;
+  double get scaleRadius => ResponsiveService.scaleRadius() * this;
 
-  double get sp => ResponsiveService.scaleText() * this;
+  double get scaleText => ResponsiveService.scaleText() * this;
 }
