+++ lib/core/networking/api_handler.dart
@@ -8,18 +8,3 @@ part 'api_handler.g.dart';
 abstract class ApiHandler {
   factory ApiHandler(Dio dio, {String baseUrl}) = _ApiHandler;
 }
-
-Dio createDioandSetupInterceptors() {
-  Dio dio = Dio();
-
-  dio.options.connectTimeout = const Duration(seconds: 20);
-  dio.options.receiveTimeout = const Duration(seconds: 20);
-
-  dio.interceptors.add(LogInterceptor(
-      request: true,
-      requestHeader: true,
-      requestBody: true,
-      responseBody: true,
-      responseHeader: true));
-  return dio;
-}
