+++ lib/core/themes/app_pallete.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
 class AppPallete {
-  static const Color backgroundColor = Color.fromRGBO(24, 24, 32, 1);
   static const Color violet = Color.fromRGBO(187, 63, 221, 1);
   static const Color pink = Color.fromRGBO(251, 109, 169, 1);
   static const Color lightOrange = Color.fromRGBO(255, 159, 124, 1);
