+++ lib/core/themes/app_theme.dart
@@ -1,6 +1,6 @@
+import 'app_pallete.dart';
 import 'package:flutter/material.dart';
 
-import 'app_pallete.dart';
 
 class AppTheme {
   static _border([Color color = AppPallete.borderColor]) => OutlineInputBorder(
@@ -8,8 +8,8 @@ class AppTheme {
       borderRadius: BorderRadius.circular(10));
 
   static final darkTthemeMode = ThemeData.dark().copyWith(
-    scaffoldBackgroundColor: AppPallete.backgroundColor,
-    appBarTheme: const AppBarTheme(backgroundColor: AppPallete.backgroundColor),
+    scaffoldBackgroundColor: AppPallete.homeBG,
+    appBarTheme: const AppBarTheme(backgroundColor: AppPallete.homeBG),
     inputDecorationTheme: InputDecorationTheme(
       contentPadding: const EdgeInsets.all(27),
       enabledBorder: _border(),
